Fixes, Indent, Improved NoCrash
Fixes: uzukings
Improved NoCrash: Creates a file ".scriptsession" on script check. It
gets deleted in serverShutDown. If the file still exists on startup,
creates a file named "nocrash.txt" and sets Config.NoCrash to true until
this file gets removed regardless of settings. Delete the file to undo
this (system crash for example)